* Fixed: correctly load RequiredMinRepValue/RequiredMaxRepValue from `quest_template` as signed values. Simplify code using this change.
git-svn-id: http://mangos.svn.sourceforge.net/svnroot/mangos@4548 26674162-90ff-0310-a85b-b64107c8dee9

+++ trunk/src/game/Player.cpp
@@ -11120,11 +11120,11 @@ bool Player::SatisfyQuestReputation( uint32 quest_id, bool msg )
     if( qInfo )
     {
         uint32 fIdMin = qInfo->GetRequiredMinRepFaction();  //Min required rep
-        if(fIdMin && GetReputation(fIdMin) < int32(qInfo->GetRequiredMinRepValue()))
+        if(fIdMin && GetReputation(fIdMin) < qInfo->GetRequiredMinRepValue())
             return false;
 
         uint32 fIdMax = qInfo->GetRequiredMaxRepFaction();  //Max required rep
-        if(fIdMax && GetReputation(fIdMax) >= int32(qInfo->GetRequiredMaxRepValue()))
+        if(fIdMax && GetReputation(fIdMax) >= qInfo->GetRequiredMaxRepValue())
             return false;
 
         return true;
